rename configs to UKCP Local pr
to make room for multivar and inform using CPM from UKCP Local (not 18)

also updated queue train helper to not require dataset now have them set to good versions in config by default
