+++ bin/queue-training
@@ -10,7 +10,7 @@ import typer
 app = typer.Typer()
 
 
-def train_cmd(sde, dataset, workdir, config, config_overrides=list):
+def train_cmd(sde, workdir, config, config_overrides=list):
     train_basecmd = ["python", f"bin/main.py"]
 
     train_opts = {
@@ -22,7 +22,6 @@ def train_cmd(sde, dataset, workdir, config, config_overrides=list):
     return (
         train_basecmd
         + [arg for item in train_opts.items() for arg in item]
-        + [f"--config.data.dataset_name={dataset}"]
         + config_overrides
     )
 
@@ -51,9 +50,8 @@ def queue_cmd(duration, memory):
 def main(
     ctx: typer.Context,
     model_run_id: str,
-    cpm_dataset: str,
     sde: str,
-    config: str = "xarray_12em_cncsnpp_continuous",
+    config: str = "ukcp_local_pr_12em_cncsnpp_continuous",
     memory: int = 64,
     duration: int = 72,
 ):
@@ -67,7 +65,7 @@ def main(
     full_cmd = (
         queue_cmd(duration=duration, memory=memory)
         + ["--"]
-        + train_cmd(sde, cpm_dataset, workdir, config, ctx.args)
+        + train_cmd(sde, workdir, config, ctx.args)
     )
     print(" ".join(full_cmd).strip(), file=sys.stderr)
     output = subprocess.run(full_cmd, capture_output=True)
