+++ src/ml_downscaling_emulator/score_sde_pytorch_hja22/configs/subvpsde/ukcp_local_pr_cunet_continuous.py
@@ -14,8 +14,8 @@
 # limitations under the License.
 
 # Lint as: python3
-"""Training NCSN++ on precip data with sub-VP SDE."""
-from ml_downscaling_emulator.score_sde_pytorch_hja22.configs.default_ukcp18_configs import get_default_configs
+"""Training conditional U-Net on precip data with sub-VP SDE."""
+from ml_downscaling_emulator.score_sde_pytorch_hja22.configs.default_ukcp_local_pr_configs import get_default_configs
 
 
 def get_config():
@@ -38,7 +38,7 @@ def get_config():
 
   # model
   model = config.model
-  model.name = 'cncsnpp'
+  model.name = 'cunet'
   model.scale_by_sigma = False
   model.ema_rate = 0.9999
   model.normalization = 'GroupNorm'
@@ -61,10 +61,5 @@ def get_config():
   model.init_scale = 0.
   model.fourier_scale = 16
   model.conv_size = 3
-
-  # data
-  data = config.data
-  data.input_transform_key = "stan"
-  data.target_transform_key = "sqrturrecen"
 
   return config
